made vaccine age float

+++ Backend/src/main/java/hackathon/medhack/backend/model/Vaccine.java
@@ -19,7 +19,7 @@ public class Vaccine {
 
     private String name;
     private String description;
-    private Integer age;
+    private float age;
     private String preventedDiseases;
     private String sideEffects;
